Added Alibaba & GoogleCloud in backup config, Fixed the schedule backup UI. Added Loader & alert

+++ angular/src/app/Storage/manage-storage/storage-detail.component.ts
@@ -16,6 +16,7 @@ import {
   BackUpStorageConfiguationServiceProxy,
   SourceConfiguationDto,
 } from "@shared/service-proxies/service-proxies";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-storage-detail",
@@ -49,6 +50,7 @@ export class StorageDetailComponent implements OnChanges, OnInit {
     this.filterData();
   }
   filterData(): void {
+    debugger;
     this.filteredEntries = this.entries.filter(
       (entry) =>
         entry.storageMasterType?.name.toLowerCase() ===
@@ -68,6 +70,10 @@ export class StorageDetailComponent implements OnChanges, OnInit {
     "nfS_AccessUserID",
     "nfS_Password",
     "nfS_LocationPath",
+    "endpoint",
+    "projectID",
+    "credentialFile",
+    "backupName",
   ];
   getFilteredKeys(entry: any): string[] {
     return Object.keys(entry).filter(
@@ -90,6 +96,10 @@ export class StorageDetailComponent implements OnChanges, OnInit {
       nfS_AccessUserID: "Access user ID",
       nfS_Password: "Password",
       nfS_LocationPath: "Location Path",
+      endpoint: "End Point",
+      projectID: "ProjectID",
+      credentialFile: "Credential File",
+      backupName: "backupName",
     };
 
     return (
@@ -115,10 +125,23 @@ export class StorageDetailComponent implements OnChanges, OnInit {
   }
 
   deleteEntryDetails(id: string) {
-    this.backUpStorageConfiguationService.delete(id).subscribe({
-      next: () => {
-        this.back.emit();
-      },
+    Swal.fire({
+      title: "Are you sure?",
+      text: "This will permanently delete the configuration!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.backUpStorageConfiguationService.delete(id).subscribe({
+          next: () => {
+            this.back.emit();
+          },
+        });
+        Swal.fire("Deleted!", "Configuration has been deleted.", "success");
+      }
     });
   }
 
@@ -144,6 +167,7 @@ export class StorageDetailComponent implements OnChanges, OnInit {
       )
       .subscribe({
         next: (result) => {
+          debugger;
           this.successfulBackupLogs = result.items || [];
           this.totalSuccessfulLogs = result.totalCount || 0;
         },
